requirements.txt
certifi==2018.10.15
chardet==3.0.4
ChatterBot==0.8.7
chatterbot-corpus==1.1.4
Click==7.0
dj-database-url==0.5.0
Django==2.1.3
django-heroku==0.3.1
Flask==1.0.2
future==0.17.1
get==2018.11.19
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
line-bot-sdk==1.8.0
MarkupSafe==1.1.0
mathparse==0.1.2
nltk==3.4
oauthlib==2.1.0
post==2018.11.20
psycopg2==2.7.6.1
public==2018.11.20
pymongo==3.7.2
python-dateutil==2.6.1
python-twitter==3.5
pytz==2018.7
PyYAML==3.13
query-string==2018.11.20
request==2018.11.20
requests==2.20.1
requests-oauthlib==1.0.0
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.14
urllib3==1.24.1
watson-developer-cloud==2.4.4
websocket-client==0.47.0
Werkzeug==0.14.1
whitenoise==4.1.2